pyproject.toml
[project]
name = "amigoai-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.27",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.0",
]
